Introduce MutationPointAddress
MutationPointAddress is introduced to store address of MutationPoint to allow MutationOperator to find that mutation point in a clone of original module.

Mutation operator now applies mutation to a module, which is given to it as argument, using the MutationPointAddress to look up for mutation point.

Value &V argument in AddMutationOperator::applyMutation is not used, and is left as deprecated for removal when it is clear it will not be needed when "remote JIT" is in place.