WIP: Possible Bug Report: the LLVM Modules and Instructions content gets corrupted before mull finishes execution
In order to proceed with finishing the work on Mutation Testing Elements support,
we need to make sure that MutationTestingElementsReporter can have access to
original LLVM Module/Instructions content that is used for generating mutation
points.

This small changeset on top of the latest master branch shows that by the
time we want to call the MutationTestingElements to generate the JSON report,
the LLVM Module/Instruction content gets corrupted.

The example demonstrates that the critical asserts of this changeset are hit
instead of going further. It is important that at the moment the Junk Detection
is run, the same mutation points works inside findAST() method, but then
it hits the assert when the findAST() is called later in the end of the driver-cxx.

This is a minimal example that reproduces the issue found during this work:

WIP: Reporting JSON to Mutation Testing Elements #517,
https://github.com/mull-project/mull/pull/517

[skip ci]