Print messages about current mutant in debug mode
Now Mull shows the same output with enable and disabled debug mode on running
mutants. It contains only progress-bar with total and a number of run mutants:

[info] Warm up run (threads: 1)
       [################################] 1/1. Finished in 26296ms
[info] Baseline run (threads: 1)
       [################################] 1/1. Finished in 25931ms
[info] Running mutants (threads: 8)
       [################################] 65704/65704. Finished in 6239680ms
[info] Mutation Testing Elements reporter: generating report to ./postgres-2021-10-02-99c5852e20.json
[info] Total execution time: 6294465ms

When an total time of mutation testing takes a lot of time it is useful to show
more information about current mutants, like file with mutant and status of
execution.